app/requirements.txt
pyyaml
pyodbc
numpy
pandas
scikit-learn
dash
waitress